requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask-SocketIO==5.1.1
pymongo==3.12.0
Flask-PyMongo==2.3.0
eventlet==0.30.2
python-dotenv==0.19.0